pyproject.toml
[project]
name = "discord-bot-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiohttp==3.9.5",
  "aiosignal==1.3.1",
  "attrs==23.2.0",
  "cachetools==5.5.2",
  "certifi==2025.1.31",
  "charset-normalizer==3.4.1",
  "discord==2.3.2",
  "discord.py==2.3.2",
  "fastapi>=0.115.12",
  "frozenlist==1.4.1",
  "google-auth==2.39.0",
  "google-auth-oauthlib==1.2.1",
  "gspread==6.2.0",
  "httplib2==0.22.0",
  "idna==3.7",
  "multidict==6.0.5",
  "oauth2client==4.1.3",
  "oauthlib==3.2.2",
  "pyasn1==0.6.1",
  "pyasn1_modules==0.4.2",
  "pyparsing==3.2.3",
  "python-dotenv==1.0.1",
  "requests==2.32.3",
  "requests-oauthlib==2.0.0",
  "rsa==4.9.1",
  "six==1.17.0",
  "sqlalchemy>=2.0.40",
  "urllib3==2.4.0",
  "uvicorn>=0.34.2",
  "yarl==1.9.4",
]
